requirements.txt
google-cloud-storage
pdfplumber
spacy
spacy-transformers